+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Change `ZestClientElementScrollTo` to only scroll to the element when not already in view.
 
 ## [0.27.0] - 2025-05-09
 ### Changed